LayoutTests/platform/mac/imported/w3c/web-platform-tests/cookies/value/value-ctl-expected.txt


PASS Cookie with %x0 in value is rejected.
PASS Cookie with %x1 in value is rejected.
PASS Cookie with %x2 in value is rejected.
PASS Cookie with %x3 in value is rejected.
PASS Cookie with %x4 in value is rejected.
PASS Cookie with %x5 in value is rejected.
PASS Cookie with %x6 in value is rejected.
PASS Cookie with %x7 in value is rejected.
PASS Cookie with %x8 in value is rejected.
PASS Cookie with %x9 in value is accepted.
FAIL Cookie with %xa in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %xb in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %xc in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %xd in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %xe in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %xf in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x10 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x11 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x12 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x13 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x14 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x15 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x16 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x17 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x18 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x19 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x1a in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x1b in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x1c in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x1d in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x1e in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x1f in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
FAIL Cookie with %x7f in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9\tvalue"
